\name{plotprotein-package}
\alias{plotprotein-package}

\docType{package}
\title{
\packageTitle{plotprotein}
}
\description{
\packageDescription{plotprotein}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{plotprotein}
\packageIndices{plotprotein}}
\author{\packageAuthor{plotprotein}
Maintainer: \packageMaintainer{plotprotein}}
\references{https://cran.r-project.org/doc/manuals/R-exts.html}
\keyword{ package }
\seealso{
code{\link{help}}
}

